Fix resources
